chat_bot/requirements.txt
environ==1.0
python-dotenv==0.19.2
sqlalchemy==1.4.29
psycopg2
alembic==1.7.5
aiogram
aioredis